DiscreteNumericalVariable is not a subclass of NumericalVariable
